chore: move MCPAPILogger to _private package
